src/bot/requirements.txt
aiogram>=3.0.0
sqlmodel>=0.0.8
python-dotenv>=1.0